Reduce renders and race conditions in policy template details (#4090)
This refactors policy template details components to reduce renders.

Additionally, fireManagedClusterView was improved to be able to
explicitly name the created ManagedClusterView objects to prevent
conflicts and properly ignore duplicate fireManagedClusterView calls.
This mostly helps local development or when the console is run in strict
mode and hooks are run twice. It also helps when two separate browser
windows are looking at the same policy template, especially Kyverno ones
with many audit violations.

Lastly, some loading states were being set to false after the
fireManagedClusterView promise was created but before it resolved. This
is now fixed to wait for the promise to be resolved.

+++ frontend/src/resources/managedclusterview.tsx
@@ -77,7 +77,9 @@ export async function fireManagedClusterView(
   resourceKind: string,
   resourceApiVersion: string,
   resourceName: string,
-  resourceNamespace?: string
+  resourceNamespace?: string,
+  managedClusterViewName?: string,
+  skipCheckExisting?: boolean
 ) {
   if (resourceKind.toLowerCase() === 'secret' || resourceKind.toLowerCase() === 'secrets') {
     // We do not allow users to view secrets as this could allow lesser permissioned users to get around RBAC.
@@ -86,40 +88,44 @@ export async function fireManagedClusterView(
         'Viewing Secrets is not allowed for security reasons. To view this secret, you must access it from the cluster directly.',
     }
   }
-  const viewName = crypto
-    .createHash('sha1')
-    .update(`${clusterName}-${resourceName}-${resourceKind}`)
-    .digest('hex')
-    .substr(0, 63)
-  // Try to get and return the managedClusterView if it exsits -> if not create one and poll
-  const getResult = await getManagedClusterView({ namespace: clusterName, name: viewName })
-    .promise.then((viewResponse) => {
-      const isProcessing = _.get(viewResponse, 'status.conditions[0].type', undefined)
-      const reason = _.get(viewResponse, 'status.conditions[0].reason', undefined)
-      const message = _.get(viewResponse, 'status.conditions[0].message', undefined)
-      if (isProcessing && reason) {
-        if (isProcessing === 'Processing' && reason === 'GetResourceProcessing') {
-          return {
-            processing: isProcessing,
-            reason: reason,
-            result: viewResponse.status?.result,
-          }
-        } else if (isProcessing === 'Processing' && reason !== 'GetResourceProcessing') {
-          return { message: message }
-        }
-      } else {
-        return {
-          message:
-            'There was an error while getting the managed resource. Make sure the managed cluster is online and healthy, and that the work manager pod in namespace open-cluster-management-agent-addon is healthy ',
-        }
-      }
-      deleteManagedClusterView({ namespace: clusterName, name: viewName })
-    })
-    .catch((err) => {
-      return err
-    })
 
-  if (getResult && getResult.code >= 400) {
+  const viewName =
+    managedClusterViewName ??
+    crypto.createHash('sha1').update(`${clusterName}-${resourceName}-${resourceKind}`).digest('hex').substr(0, 63)
+
+  let getResult: any
+
+  if (!skipCheckExisting) {
+    // Try to get and return the managedClusterView if it exsits -> if not create one and poll
+    getResult = await getManagedClusterView({ namespace: clusterName, name: viewName })
+      .promise.then((viewResponse) => {
+        const isProcessing = _.get(viewResponse, 'status.conditions[0].type', undefined)
+        const reason = _.get(viewResponse, 'status.conditions[0].reason', undefined)
+        const message = _.get(viewResponse, 'status.conditions[0].message', undefined)
+        if (isProcessing && reason) {
+          if (isProcessing === 'Processing' && reason === 'GetResourceProcessing') {
+            return {
+              processing: isProcessing,
+              reason: reason,
+              result: viewResponse.status?.result,
+            }
+          } else if (isProcessing === 'Processing' && reason !== 'GetResourceProcessing') {
+            return { message: message }
+          }
+        } else {
+          return {
+            message:
+              'There was an error while getting the managed resource. Make sure the managed cluster is online and healthy, and that the work manager pod in namespace open-cluster-management-agent-addon is healthy ',
+          }
+        }
+        deleteManagedClusterView({ namespace: clusterName, name: viewName })
+      })
+      .catch((err) => {
+        return err
+      })
+  }
+
+  if (!getResult || getResult.code >= 400) {
     const { apiGroup, version } = getGroupFromApiVersion(resourceApiVersion)
     const body: ManagedClusterView = {
       apiVersion: ManagedClusterViewApiVersion,
@@ -145,13 +151,6 @@ export async function fireManagedClusterView(
     return createResource<ManagedClusterView>(body)
       .promise.then(async () => {
         return pollManagedClusterView(viewName, clusterName)
-      })
-      .catch(async (err) => {
-        if (err instanceof ResourceError && err.reason === 'AlreadyExists') {
-          return pollManagedClusterView(viewName, clusterName)
-        } else {
-          throw err
-        }
       })
       .catch((err) => {
         console.error(err)
